+++ core/src/test/java/eu/bittrade/libs/steemj/base/models/operations/virtual/ProducerRewardOperationIT.java
@@ -4,6 +4,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.junit.BeforeClass;
@@ -27,7 +28,7 @@ public class ProducerRewardOperationIT extends BaseITForOperationParsing {
     private static final int OPERATION_INDEX = 0;
     private static final String EXPECTED_PRODUCER = "xeldal";
     private static final AssetSymbolType EXPECTED_VESTS_SYMBOL = AssetSymbolType.VESTS;
-    private static final double EXPECTED_VESTS_VALUE_REAL = 390.97665;
+    private static final BigDecimal EXPECTED_VESTS_VALUE_REAL = BigDecimal.valueOf(390.97665);
     private static final long EXPECTED_VESTS_VALUE = 390976650L;
 
     /**
